v0.2.1 - Add remainder of skill tables

+++ charclass/models.py
@@ -3,8 +3,8 @@ from django.db import models
 
 class CharClass(models.Model):
     name = models.CharField(unique=True, max_length=32)
-    base_body = models.IntegerField()
-    max_armour = models.IntegerField()
+    base_body = models.PositiveSmallIntegerField()
+    max_armour = models.PositiveSmallIntegerField()
     description = models.TextField()
 
     class Meta:
